handlers: added prototype swupd handler.
Added a prototype handler for enabling swupd-based
image updates in swupdate. swupd is the update
utility primarily developed for and used by Clear
Linux. For more information see the introduction at

  https://clearlinux.org/features/software-update

The prototype handler does not imeplement any logic
as such. It merely extracts and exposes the supplied
configuration to a wrapper script which is expected
to take care of invoking swupd for the update. The
handler also proxies progress and error messages
back from the update process to the swupdate logging
infrastructure.

The wrapper script is a collection of scriptlets. Ideally
each scriptlet takes care of a single well-defined aspect
of the update process. The scriptlets are run in the same
process context so they can share/pass on information to
each other. The update process is considered to have failed
if any of the scriptlets fails.

Initial hook stubs coming in a subsequent commit.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>