handlers: added a proto wrapper handler for generic update backends.
Added an alternative way of wrapping existing software updaters into
swupdate. The handler itself does not implement almost any logic. It
merely extracts and exposes the supplied configuration to a set of
backend-specific scriptlets/hooks as environment variables. It then
runs the backend hooks until sucessful completion or an error. The
handler also proxies progress and error messages back from the update
process to the swupdate logging infrastructure.

The backend is selected by setting the image type to the name of the
backend. Also the top-level configuration node in the configuration
file supplied as part of the .swu must be named by the backend. The
backends are dynamically discovered and registered upon startup by
scanning a well-known directory for subdirectories (which are used
to store backend-specific scripts).

A wrapper/wrapped backend at the swupdate-level is practically a set
of scriptlets, called hooks, which collectively handle all the update
logic for that particular backend. Ideally each hook should take care
of a single well-defined aspect of the update process (allowing for
easier sharing of hooks among different backends) but nothing prevents
one from handling everything in a single hook script.

The hooks are all run by default in the same context (as the main hook)
thus they can easily share data by setting well-defined variables. The
hooks are also provided with a small set of shell functions for taking
care of logging informational, warning, and error messages, among other
things.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>