swupdate: added a Makefile.install for providing install targets.
While the customised Kbuild-based build system provides most of
the usual / expected targets it does not provide an install target.
This commit adds a simple bare-bone Makefile.install which provides
targets for installing the swupdate binary and installing the main
wrapper hook with any selected backend hooks that might be part of
the tree.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>